+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 - `core` library that implements all the features that existed in rhai `std`.
 - `Script` node for running script periodically or on event.
+- `ipc_request` plugin API for calling ewwii IPC directly.
 
 ### Changed